requirements.txt
streamlit
pandas
numpy
plotly
requests
pytrends
python-dotenv
pycoingecko